LPS-119068 Avoid run dispatch of the portlet.
