+++ modules/apps/layout/layout-reports-web/src/main/java/com/liferay/layout/reports/web/internal/product/navigation/control/menu/LayoutReportsProductNavigationControlMenuEntry.java
@@ -14,10 +14,19 @@
 
 package com.liferay.layout.reports.web.internal.product.navigation.control.menu;
 
+import com.liferay.configuration.admin.constants.ConfigurationAdminPortletKeys;
+import com.liferay.frontend.js.loader.modules.extender.npm.NPMResolver;
+import com.liferay.info.constants.InfoDisplayWebKeys;
+import com.liferay.info.item.InfoItemDetails;
+import com.liferay.info.item.InfoItemServiceTracker;
+import com.liferay.info.item.provider.InfoItemFieldValuesProvider;
+import com.liferay.layout.reports.web.internal.configuration.LayoutReportsGooglePageSpeedCompanyConfiguration;
 import com.liferay.layout.reports.web.internal.configuration.LayoutReportsGooglePageSpeedConfiguration;
 import com.liferay.layout.reports.web.internal.configuration.provider.LayoutReportsGooglePageSpeedConfigurationProvider;
 import com.liferay.layout.reports.web.internal.constants.LayoutReportsPortletKeys;
-import com.liferay.petra.portlet.url.builder.PortletURLBuilder;
+import com.liferay.layout.reports.web.internal.data.provider.LayoutReportsDataProvider;
+import com.liferay.layout.seo.kernel.LayoutSEOLink;
+import com.liferay.layout.seo.kernel.LayoutSEOLinkManager;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
@@ -25,9 +34,9 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.module.configuration.ConfigurationProvider;
-import com.liferay.portal.kernel.portlet.LiferayWindowState;
 import com.liferay.portal.kernel.portlet.PortalPreferences;
 import com.liferay.portal.kernel.portlet.PortletPreferencesFactoryUtil;
 import com.liferay.portal.kernel.portlet.PortletURLFactory;
@@ -40,7 +49,9 @@ import com.liferay.portal.kernel.service.permission.LayoutPermissionUtil;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.Html;
+import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.PropsKeys;
@@ -48,26 +59,31 @@ import com.liferay.portal.kernel.util.PropsUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.portal.kernel.util.SessionClicks;
 import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.portal.template.react.renderer.ComponentDescriptor;
+import com.liferay.portal.template.react.renderer.ReactRenderer;
 import com.liferay.product.navigation.control.menu.BaseProductNavigationControlMenuEntry;
 import com.liferay.product.navigation.control.menu.ProductNavigationControlMenuEntry;
 import com.liferay.product.navigation.control.menu.constants.ProductNavigationControlMenuCategoryKeys;
 import com.liferay.taglib.aui.IconTag;
-import com.liferay.taglib.portletext.RuntimeTag;
 import com.liferay.taglib.util.BodyBottomTag;
 
 import java.io.IOException;
 import java.io.Writer;
 
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.ResourceBundle;
+import java.util.stream.Collectors;
 
-import javax.portlet.RenderRequest;
-import javax.portlet.WindowStateException;
+import javax.portlet.PortletRequest;
+import javax.portlet.PortletURL;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -149,20 +165,9 @@ public class LayoutReportsProductNavigationControlMenuEntry
 
 		if (_isPanelStateOpen(httpServletRequest)) {
 			values.put("cssClass", "active");
-			values.put("layoutReportsPanelURL", StringPool.BLANK);
 		}
 		else {
 			values.put("cssClass", StringPool.BLANK);
-
-			try {
-				values.put(
-					"layoutReportsPanelURL",
-					_getLayoutReportsPanelURL(
-						httpServletRequest, _portal, _portletURLFactory));
-			}
-			catch (WindowStateException windowStateException) {
-				throw new IOException(windowStateException);
-			}
 		}
 
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
@@ -242,22 +247,240 @@ public class LayoutReportsProductNavigationControlMenuEntry
 			LayoutReportsPortletKeys.LAYOUT_REPORTS);
 	}
 
-	private String _getLayoutReportsPanelURL(
-			HttpServletRequest httpServletRequest, Portal portal,
-			PortletURLFactory portletURLFactory)
-		throws WindowStateException {
+	private Map<Locale, String> _getAlternateURLs(
+		String currentCompleteURL, Layout layout, ThemeDisplay themeDisplay) {
 
-		return PortletURLBuilder.create(
-			portletURLFactory.create(
-				httpServletRequest, LayoutReportsPortletKeys.LAYOUT_REPORTS,
-				RenderRequest.RENDER_PHASE)
-		).setMVCRenderCommandName(
-			"/layout_reports/layout_reports_panel"
-		).setRedirect(
-			portal.getCurrentCompleteURL(httpServletRequest)
-		).setWindowState(
-			LiferayWindowState.EXCLUSIVE
-		).buildString();
+		try {
+			return _portal.getAlternateURLs(
+				currentCompleteURL, themeDisplay, layout);
+		}
+		catch (PortalException portalException) {
+			_log.error(portalException, portalException);
+		}
+
+		return Collections.emptyMap();
+	}
+
+	private String _getCanonicalURL(
+		String currentCompleteURL, Layout layout, ThemeDisplay themeDisplay) {
+
+		try {
+			return _portal.getCanonicalURL(
+				currentCompleteURL, themeDisplay, layout, false, false);
+		}
+		catch (PortalException portalException) {
+			_log.error(portalException, portalException);
+		}
+
+		return StringPool.BLANK;
+	}
+
+	private List<Map<String, Object>> _getCanonicalURLs(
+		HttpServletRequest httpServletRequest, Layout layout,
+		ThemeDisplay themeDisplay) {
+
+		Locale defaultLocale = _getDefaultLocale(layout);
+
+		String canonicalURL = _getCanonicalURL(
+			_getCompleteURL(httpServletRequest, themeDisplay), layout,
+			themeDisplay);
+
+		Map<Locale, String> alternateURLs = _getAlternateURLs(
+			canonicalURL, layout, themeDisplay);
+
+		return Optional.ofNullable(
+			_groupLocalService.fetchGroup(layout.getGroupId())
+		).map(
+			Group::getGroupId
+		).map(
+			_language::getAvailableLocales
+		).orElseGet(
+			Collections::emptySet
+		).stream(
+		).sorted(
+			(locale1, locale2) -> {
+				if (Objects.equals(locale1, defaultLocale)) {
+					return -1;
+				}
+
+				if (Objects.equals(locale2, defaultLocale)) {
+					return 1;
+				}
+
+				String languageId1 = LocaleUtil.toW3cLanguageId(locale1);
+				String languageId2 = LocaleUtil.toW3cLanguageId(locale2);
+
+				return languageId1.compareToIgnoreCase(languageId2);
+			}
+		).map(
+			locale -> HashMapBuilder.<String, Object>put(
+				"canonicalURL",
+				() -> {
+					LayoutSEOLink layoutSEOLink =
+						_layoutSEOLinkManager.getCanonicalLayoutSEOLink(
+							layout, locale, canonicalURL, alternateURLs);
+
+					return layoutSEOLink.getHref();
+				}
+			).put(
+				"languageId", LocaleUtil.toW3cLanguageId(locale)
+			).put(
+				"title", _getTitle(httpServletRequest, layout, locale)
+			).build()
+		).collect(
+			Collectors.toList()
+		);
+	}
+
+	private String _getCompleteURL(
+		HttpServletRequest httpServletRequest, ThemeDisplay themeDisplay) {
+
+		try {
+			return _portal.getLayoutURL(themeDisplay);
+		}
+		catch (PortalException portalException) {
+			_log.error(portalException, portalException);
+
+			return _portal.getCurrentCompleteURL(httpServletRequest);
+		}
+	}
+
+	private String _getConfigureGooglePageSpeedURL(
+		HttpServletRequest httpServletRequest, ThemeDisplay themeDisplay) {
+
+		if (_isOmniAdmin()) {
+			PortletURL portletURL = _portal.getControlPanelPortletURL(
+				httpServletRequest,
+				ConfigurationAdminPortletKeys.SYSTEM_SETTINGS,
+				PortletRequest.RENDER_PHASE);
+
+			portletURL.setParameter(
+				"mvcRenderCommandName",
+				"/configuration_admin/edit_configuration");
+			portletURL.setParameter(
+				"redirect", _portal.getCurrentCompleteURL(httpServletRequest));
+			portletURL.setParameter(
+				"factoryPid",
+				LayoutReportsGooglePageSpeedConfiguration.class.getName());
+			portletURL.setParameter(
+				"pid",
+				LayoutReportsGooglePageSpeedConfiguration.class.getName());
+
+			return portletURL.toString();
+		}
+		else if (_isCompanyAdmin()) {
+			PortletURL portletURL = _portal.getControlPanelPortletURL(
+				httpServletRequest,
+				ConfigurationAdminPortletKeys.INSTANCE_SETTINGS,
+				PortletRequest.RENDER_PHASE);
+
+			portletURL.setParameter(
+				"mvcRenderCommandName",
+				"/configuration_admin/edit_configuration");
+			portletURL.setParameter(
+				"redirect", _portal.getCurrentCompleteURL(httpServletRequest));
+			portletURL.setParameter(
+				"factoryPid",
+				LayoutReportsGooglePageSpeedCompanyConfiguration.class.
+					getName());
+			portletURL.setParameter(
+				"pid",
+				LayoutReportsGooglePageSpeedCompanyConfiguration.class.
+					getName());
+
+			return portletURL.toString();
+		}
+		else if (_isSiteAdmin(themeDisplay.getScopeGroupId())) {
+			try {
+				PortletURL portletURL = _portal.getControlPanelPortletURL(
+					httpServletRequest,
+					_groupLocalService.getGroup(themeDisplay.getScopeGroupId()),
+					"com_liferay_site_admin_web_portlet_SiteSettingsPortlet", 0,
+					0, PortletRequest.RENDER_PHASE);
+
+				return portletURL.toString();
+			}
+			catch (PortalException portalException) {
+				_log.error(portalException, portalException);
+
+				return null;
+			}
+		}
+
+		return null;
+	}
+
+	private Map<String, Object> _getData(
+		HttpServletRequest httpServletRequest, Layout layout,
+		LayoutReportsDataProvider layoutReportsDataProvider,
+		ThemeDisplay themeDisplay) {
+
+		return Collections.singletonMap(
+			"context",
+			HashMapBuilder.<String, Object>put(
+				"assetsPath",
+				_portal.getPathContext(httpServletRequest) + "/assets/"
+			).put(
+				"canonicalURLs",
+				_getCanonicalURLs(httpServletRequest, layout, themeDisplay)
+			).put(
+				"configureGooglePageSpeedURL",
+				_getConfigureGooglePageSpeedURL(
+					httpServletRequest, themeDisplay)
+			).put(
+				"defaultLanguageId",
+				LocaleUtil.toW3cLanguageId(_getDefaultLocale(layout))
+			).put(
+				"validConnection", layoutReportsDataProvider.isValidConnection()
+			).build());
+	}
+
+	private Locale _getDefaultLocale(Layout layout) {
+		try {
+			return _portal.getSiteDefaultLocale(layout.getGroupId());
+		}
+		catch (PortalException portalException) {
+			_log.error(portalException, portalException);
+
+			return LocaleUtil.getSiteDefault();
+		}
+	}
+
+	private String _getTitle(
+		HttpServletRequest httpServletRequest, Layout layout, Locale locale) {
+
+		if (layout.isTypeAssetDisplay()) {
+			return Optional.ofNullable(
+				(InfoItemDetails)httpServletRequest.getAttribute(
+					InfoDisplayWebKeys.INFO_ITEM_DETAILS)
+			).map(
+				infoItemDetails ->
+					_infoItemServiceTracker.getFirstInfoItemService(
+						InfoItemFieldValuesProvider.class,
+						infoItemDetails.getClassName())
+			).map(
+				infoItemFieldValuesProvider ->
+					infoItemFieldValuesProvider.getInfoItemFieldValue(
+						httpServletRequest.getAttribute(
+							InfoDisplayWebKeys.INFO_ITEM),
+						"title")
+			).map(
+				infoFieldValue -> (String)infoFieldValue.getValue(locale)
+			).orElse(
+				StringPool.BLANK
+			);
+		}
+		else if (layout.isTypeContent() || layout.isTypePortlet()) {
+			return Optional.ofNullable(
+				layout.getTitle(locale)
+			).filter(
+				Validator::isNotNull
+			).orElseGet(
+				() -> layout.getName(locale)
+			);
+		}
+
+		return StringPool.BLANK;
 	}
 
 	private boolean _hasViewPermission(
@@ -271,6 +494,13 @@ public class LayoutReportsProductNavigationControlMenuEntry
 		}
 
 		return true;
+	}
+
+	private boolean _isCompanyAdmin() {
+		PermissionChecker permissionChecker =
+			PermissionThreadLocal.getPermissionChecker();
+
+		return permissionChecker.isCompanyAdmin();
 	}
 
 	private boolean _isEmbeddedPersonalApplicationLayout(Layout layout) {
@@ -288,6 +518,13 @@ public class LayoutReportsProductNavigationControlMenuEntry
 		}
 
 		return false;
+	}
+
+	private boolean _isOmniAdmin() {
+		PermissionChecker permissionChecker =
+			PermissionThreadLocal.getPermissionChecker();
+
+		return permissionChecker.isOmniadmin();
 	}
 
 	private boolean _isPanelStateOpen(HttpServletRequest httpServletRequest) {
@@ -349,6 +586,13 @@ public class LayoutReportsProductNavigationControlMenuEntry
 		return true;
 	}
 
+	private boolean _isSiteAdmin(long groupId) {
+		PermissionChecker permissionChecker =
+			PermissionThreadLocal.getPermissionChecker();
+
+		return permissionChecker.isGroupAdmin(groupId);
+	}
+
 	private void _processBodyBottomTagBody(PageContext pageContext)
 		throws IOException, JspException {
 
@@ -378,11 +622,33 @@ public class LayoutReportsProductNavigationControlMenuEntry
 		jspWriter.write(
 			"<div class=\"sidebar sidebar-light sidenav-menu sidebar-sm\">");
 
-		RuntimeTag runtimeTag = new RuntimeTag();
+		ThemeDisplay themeDisplay =
+			(ThemeDisplay)httpServletRequest.getAttribute(
+				WebKeys.THEME_DISPLAY);
 
-		runtimeTag.setPortletName(LayoutReportsPortletKeys.LAYOUT_REPORTS);
+		Layout layout = _layoutLocalService.fetchLayout(themeDisplay.getPlid());
 
-		runtimeTag.doTag(pageContext);
+		try {
+			Group group = _groupLocalService.getGroup(
+				themeDisplay.getScopeGroupId());
+
+			LayoutReportsDataProvider layoutReportsDataProvider =
+				new LayoutReportsDataProvider(
+					_layoutReportsGooglePageSpeedConfigurationProvider.
+						getApiKey(group));
+
+			_reactRenderer.renderReact(
+				new ComponentDescriptor(
+					_npmResolver.resolveModuleName("layout-reports-web") +
+						"/js/LayoutReportsApp"),
+				_getData(
+					httpServletRequest, layout, layoutReportsDataProvider,
+					themeDisplay),
+				httpServletRequest, jspWriter);
+		}
+		catch (Exception exception) {
+			throw new IOException(exception);
+		}
 
 		jspWriter.write("</div></div>");
 	}
@@ -403,6 +669,9 @@ public class LayoutReportsProductNavigationControlMenuEntry
 	private Html _html;
 
 	@Reference
+	private InfoItemServiceTracker _infoItemServiceTracker;
+
+	@Reference
 	private Language _language;
 
 	@Reference
@@ -412,11 +681,20 @@ public class LayoutReportsProductNavigationControlMenuEntry
 		_layoutReportsGooglePageSpeedConfigurationProvider;
 
 	@Reference
+	private LayoutSEOLinkManager _layoutSEOLinkManager;
+
+	@Reference
+	private NPMResolver _npmResolver;
+
+	@Reference
 	private Portal _portal;
 
 	private String _portletNamespace;
 
 	@Reference
 	private PortletURLFactory _portletURLFactory;
+
+	@Reference
+	private ReactRenderer _reactRenderer;
 
 }
