+++ modules/apps/layout/layout-reports-web/src/main/resources/META-INF/resources/init.jsp
@@ -17,16 +17,10 @@
 <%@ taglib uri="http://java.sun.com/portlet_2_0" prefix="portlet" %>
 
 <%@ taglib uri="http://liferay.com/tld/aui" prefix="aui" %><%@
-taglib uri="http://liferay.com/tld/clay" prefix="clay" %><%@
-taglib uri="http://liferay.com/tld/react" prefix="react" %><%@
-taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
-taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %><%@
-taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
+taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %>
 
 <liferay-theme:defineObjects />
 
-<%@ page import="com.liferay.layout.reports.web.internal.constants.LayoutReportsWebKeys" %><%@
-page import="com.liferay.layout.reports.web.internal.display.context.LayoutReportsDisplayContext" %><%@
-page import="com.liferay.layout.reports.web.internal.display.context.LayoutReportsGooglePageSpeedDisplayContext" %>
+<%@ page import="com.liferay.layout.reports.web.internal.display.context.LayoutReportsGooglePageSpeedDisplayContext" %>
 
 <%@ include file="/init-ext.jsp" %>
