--- modules/apps/layout/layout-reports-web/src/main/resources/META-INF/resources/layout_reports_panel.jsp
@@ -1,32 +1,0 @@
-<%--
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
---%>
-
-<%@ include file="/init.jsp" %>
-
-<%
-LayoutReportsDisplayContext layoutReportsDisplayContext = (LayoutReportsDisplayContext)request.getAttribute(LayoutReportsWebKeys.LAYOUT_REPORTS_DISPLAY_CONTEXT);
-%>
-
-<div class="c-p-3" id="<portlet:namespace />-layout-reports-root">
-	<div class="inline-item my-5 p-5 w-100">
-		<span aria-hidden="true" class="loading-animation"></span>
-	</div>
-
-	<react:component
-		module="js/LayoutReportsApp"
-		props="<%= layoutReportsDisplayContext.getData() %>"
-	/>
-</div>
